Move route geometry validation to own hook
