+++ src/hooks/routes/useCreateRoute.ts
@@ -1,32 +1,22 @@
 import { useTranslation } from 'react-i18next';
 import { RouteFormState } from '../../components/forms/route/RoutePropertiesForm.types';
 import {
-  GetStopsByIdsDocument,
-  GetStopsByIdsQuery,
-  GetStopsByIdsQueryVariables,
   InsertRouteOneMutationVariables,
   RouteRoute,
   useInsertRouteOneMutation,
 } from '../../generated/graphql';
-import {
-  InfrastructureLinkAlongRoute,
-  mapGetStopsResult,
-  mapInfraLinksAlongRouteToGraphQL,
-} from '../../graphql';
+import { mapInfraLinksAlongRouteToGraphQL, RouteGeometry } from '../../graphql';
 import { MIN_DATE } from '../../time';
 import {
   mapToObject,
   mapToVariables,
   showDangerToastWithError,
 } from '../../utils';
-import { useAsyncQuery } from '../useAsyncQuery';
 import { useCheckValidityAndPriorityConflicts } from '../useCheckValidityAndPriorityConflicts';
 import { mapRouteFormToInput } from './useEditRouteMetadata';
+import { useValidateRoute } from './useValidateRoute';
+import { extractFirstAndLastStopFromStops } from './utils';
 
-export interface RouteGeometry {
-  stopIdsWithinRoute: UUID[];
-  infraLinksAlongRoute: InfrastructureLinkAlongRoute[];
-}
 interface CreateParams {
   form: RouteFormState;
   routeGeometry: RouteGeometry;
@@ -47,73 +37,11 @@ export const mapStopsToStopSequence = (stops: UUID[]) => {
   };
 };
 
-export const extractFirstAndLastStopFromStops = (
-  stopIdsWithinRoute: UUID[],
-) => {
-  // TODO: These will be removed from schema, remove from here as well
-  const startingStopId = stopIdsWithinRoute[0];
-  const finalStopId = stopIdsWithinRoute[stopIdsWithinRoute.length - 1];
-
-  return { startingStopId, finalStopId };
-};
-
 export const useCreateRoute = () => {
   const { t } = useTranslation();
   const [mutateFunction] = useInsertRouteOneMutation();
   const { getConflictingRoutes } = useCheckValidityAndPriorityConflicts();
-  const [getStopsByIds] = useAsyncQuery<
-    GetStopsByIdsQuery,
-    GetStopsByIdsQueryVariables
-  >(GetStopsByIdsDocument);
-
-  const validateGeometry = async ({
-    stopIdsWithinRoute,
-    infraLinksAlongRoute,
-  }: RouteGeometry) => {
-    // Check that there are enoung stops on the route
-
-    if (
-      !infraLinksAlongRoute ||
-      !stopIdsWithinRoute ||
-      stopIdsWithinRoute.length < 2
-    ) {
-      throw new Error(t('routes.tooFewStops'));
-    }
-
-    // Check that route's starting stop resides on starting infrastructure link
-    // and route's final stop resides on final infrastructure link
-
-    const { startingStopId, finalStopId } =
-      extractFirstAndLastStopFromStops(stopIdsWithinRoute);
-
-    const stopsResult = await getStopsByIds({
-      stopIds: [startingStopId, finalStopId],
-    });
-
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const stops = mapGetStopsResult(stopsResult)!;
-
-    const startingStop = stops[0];
-    const finalStop = stops[1];
-
-    const startingInfraLink = infraLinksAlongRoute[0];
-    const finalInfraLink =
-      infraLinksAlongRoute[infraLinksAlongRoute.length - 1];
-
-    if (
-      startingStop.located_on_infrastructure_link_id !==
-      startingInfraLink.infrastructureLinkId
-    ) {
-      throw new Error(t('routes.startingStopNotOnStartingInfraLink'));
-    }
-
-    if (
-      finalStop.located_on_infrastructure_link_id !==
-      finalInfraLink.infrastructureLinkId
-    ) {
-      throw new Error(t('routes.finalStopNotOnFinalInfraLink'));
-    }
-  };
+  const { validateGeometry } = useValidateRoute();
 
   const mapRouteDetailsToInsertMutationVariables = (
     params: CreateParams,
@@ -180,6 +108,5 @@ export const useCreateRoute = () => {
     mapCreateChangesToVariables,
     insertRouteMutation: mutateFunction,
     defaultErrorHandler,
-    validateGeometry,
   };
 };
